LPS-202074 Move LockedLayoutType to layout-api
